refactor: unify feature navigation
